refactor navigation into single component
